refactor: 모듈에 절대경로 지정
