+++ client/src/components/Chat/ChatItem/index.tsx
@@ -2,8 +2,8 @@ import React, { useState } from 'react';
 import { useDispatch } from 'react-redux';
 import { postLikeChat } from '../../../api/postLikeChat';
 import { STATUS_CODES } from '../../../api/STATUS_CODES';
-import { setSelectedChat } from '../../../redux/selectedChat/slice';
-import { ChatData } from '../../../types/chats';
+import { setSelectedChat } from '@redux/selectedChat/slice';
+import { ChatData } from '@customTypes/chats';
 import ThreadPreview from '../ThreadPreview';
 import AddChatReaction from '../AddChatReaction';
 import ChatReaction from '../ChatReaction';
