+++ client/src/components/Chat/Thread/index.tsx
@@ -3,9 +3,9 @@ import { useDispatch } from 'react-redux';
 import useSWR from 'swr';
 import { API_URL } from '../../../api/API_URL';
 import { postCreateThread } from '../../../api/postCreateThread';
-import { useSelectedChannel } from '../../../hooks/useSelectedChannel';
-import { setSelectedChat } from '../../../redux/selectedChat/slice';
-import { ChatData } from '../../../types/chats';
+import { useSelectedChannel } from '@hooks/useSelectedChannel';
+import { setSelectedChat } from '@redux/selectedChat/slice';
+import { ChatData } from '@customTypes/chats';
 import { getFetcher } from '../../../util/fetcher';
 import { socket } from '../../../util/socket';
 import ThreadItem from '../ThreadItem';
