+++ client/src/components/Meet/MeetChat/index.tsx
@@ -1,6 +1,6 @@
 import React, { FormEvent, useEffect, useLayoutEffect, useRef, useState } from 'react';
-import { useSelectedChannel } from '../../../hooks/useSelectedChannel';
-import { useUserdata } from '../../../hooks/useUserdata';
+import { useSelectedChannel } from '@hooks/useSelectedChannel';
+import { useUserdata } from '@hooks/useUserdata';
 import { socket } from '../../../util/socket';
 import { ChatHeader, UserImage } from '../../Chat/ChatItem/style';
 import {
