+++ client/src/components/Meet/MeetVideo/index.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect, useRef, useCallback } from 'react';
-import { useSelectedChannel } from '../../../hooks/useSelectedChannel';
-import { useUserdata } from '../../../hooks/useUserdata';
-import { useUserDevice } from '../../../hooks/useUserDevice';
+import { useSelectedChannel } from '@hooks/useSelectedChannel';
+import { useUserdata } from '@hooks/useUserdata';
+import { useUserDevice } from '@hooks/useUserDevice';
 import Socket, { socket } from '../../../util/socket';
 import { MeetVideoWrapper, VideoItemWrapper, VideoItem } from './style';
