+++ client/src/components/Modal/GroupJoin/index.tsx
@@ -3,10 +3,10 @@ import { useDispatch } from 'react-redux';
 import { useHistory } from 'react-router';
 import Modal from '..';
 import { postJoinGroup } from '../../../api/postJoinGroup';
-import { useGroups } from '../../../hooks/useGroups';
-import { setSelectedGroup } from '../../../redux/selectedGroup/slice';
-import Colors from '../../../styles/Colors';
-import { ModalController } from '../../../types/modal';
+import { useGroups } from '@hooks/useGroups';
+import { setSelectedGroup } from '@redux/selectedGroup/slice';
+import Colors from '@styles/Colors';
+import { ModalController } from '@customTypes/modal';
 import { Input } from './style';
 
 function GroupJoinModal({ controller: { hide, show } }: { controller: ModalController }) {
