+++ client/src/hooks/useGroups.ts
@@ -2,7 +2,7 @@ import { useEffect, useLayoutEffect } from 'react';
 import { useDispatch } from 'react-redux';
 import useSWR from 'swr';
 import { API_URL } from '../api/API_URL';
-import { setSelectedGroup } from '../redux/selectedGroup/slice';
+import { setSelectedGroup } from '@redux/selectedGroup/slice';
 import { getFetcher } from '../util/fetcher';
 import { socket } from '../util/socket';
 import { useSelectedGroup } from './useSelectedGroup';
