+++ client/src/hooks/useSelectedGroup.ts
@@ -1,4 +1,4 @@
 import { useSelector } from 'react-redux';
-import { RootState } from '../redux/store';
+import { RootState } from '@redux/store';
 
 export const useSelectedGroup = () => useSelector((state: RootState) => state.selectedGroup);
