+++ client/src/index.tsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
 
-import store from './redux/store';
+import store from '@redux/store';
 import App from './App';
 
 ReactDOM.render(
