+++ client/src/pages/Main/index.tsx
@@ -1,14 +1,14 @@
 import React, { useLayoutEffect } from 'react';
 import { useDispatch } from 'react-redux';
-import ChannelHeader from '../../components/ChannelHeader';
-import Chat from '../../components/Chat';
-import Meet from '../../components/Meet';
-import SideBar from '../../components/SideBar';
-import { useGroups } from '../../hooks/useGroups';
-import { useSelectedChannel } from '../../hooks/useSelectedChannel';
-import { useSelectedGroup } from '../../hooks/useSelectedGroup';
-import { setSelectedChannel } from '../../redux/selectedChannel/slice';
-import { setSelectedGroup } from '../../redux/selectedGroup/slice';
+import ChannelHeader from '@components/ChannelHeader';
+import Chat from '@components/Chat';
+import Meet from '@components/Meet';
+import SideBar from '@components/SideBar';
+import { useGroups } from '@hooks/useGroups';
+import { useSelectedChannel } from '@hooks/useSelectedChannel';
+import { useSelectedGroup } from '@hooks/useSelectedGroup';
+import { setSelectedChannel } from '@redux/selectedChannel/slice';
+import { setSelectedGroup } from '@redux/selectedGroup/slice';
 import { getURLParams } from '../../util/getURLParams';
 import { Layout, MainWrapper } from './style';
